backend/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.3.0
certifi==2025.6.15
cffi==1.17.1
click==8.1.8
cryptography==45.0.4
dnspython==2.6.1
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.13
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.15
packaging==25.0
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.5
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.2
pytest==8.3.5
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.44.0
tomli==2.2.1
typer==0.16.0
typing_extensions==4.13.2
ujson==5.10.0
uvicorn==0.33.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
zipp==3.20.2
